Test shift counts for int8_t and int16_t that are too large w/ explicit masks
The SPIR-V spec says:

    The result is undefined if Shift is greater than or equal to the bit
    width of the components of Base.

However, other APIs specify that the shift count is modulo the bitwidth
of Base.  To achieve this in SPIR-V, translators will insert explicit
masks of Shift with 7, 15, 31, or 63 depending on the type of Base.

These two tests verify that these explicit masks are respected for 8-bit
and 16-bit Base.

The Intel PRMs imply that on BDW and earlier, Shift is implicitly
masked with 31 when Base is 32-bits or smaller.  Optimizations in Mesa
remove explicit masks that match the size of Base (e.g., x << (y & 15)
for 16-bit Base) with the assumption that the native instruction will
implicitly do this mask.

The Intel PRMs explicitly say that on SKL and later, Shift is
implicitly masked with the proper mask for the size of Base.

The results of this test seem to contradict the PRMs.  If the PRMs are
correct, the test should pass on every SKL+ platform, and it should fail
on BDW.  Earlier platforms either don't support shaderInt16 or
shaderInt8, or don't support Vulkan.  However, the total list of
failures is:

    crucible.func.shader.shift.int8_t.q0.gen9atomm64
    crucible.func.shader.shift.int8_t.q0.bswm64
    crucible.func.shader.shift.int8_t.q0.bdwm64
    crucible.func.shader.shift.int8_t.q0.gen9m64

The 16-bit test passes on BDW, and the 8-bit test fails on BDW and
SKL.

v2: Update include path.  Update file name to reflect that it contains
both Int8 and Int16 tests.  Suggested by Caio.

Reviewed-by: Caio Marcelo de Oliveira Filho <caio.oliveira@intel.com>