shaderInt8Int16-shift: Ensure the shift counts will produce the failure
In 7d03ddc, the shift counts for int16_t did not have 0x10 set, so the
bug would not be reproduced. With this change to the test, I have
verified that func.shader.shift.int16_t fails on BDW and ICL. Based on
that, I assume it will fail on other platforms as well.

Fixes: 7d03ddc ("Test shift counts for int8_t and int16_t that are too large w/ explicit masks")