dispatcher: Poll regularly if a test timeout is set
Normally, if all the workers are busy, we wait "forever" for new
results. Also, when cleaning up results at the end, we will wait
"forever" for tests to finish.

If a timeout is set, then this will cause the timeout to no be
observed. Therefore, we now wait for 1/10th of the timeout time in
these cases.

If a timeout is not set, then we will wait "forever" as before.

Signed-off-by: Jordan Justen <jordan.l.justen@intel.com>