miptree: Don't discard 3D slices on intermediate copy image
For the intermediate step of copying to an image, the test has been
performing a layout transition from UNDEFINED for each depth slice of 3D
intermediate images. This is redundant because a layout transition on a
3D image already occurs for all depth slices of a given miplevel.

This would have simply been an inefficiency if the transitions were
batched one after another. However, between each transition of the
intermediate image, the test copied data to that same image. As a
result, each transition after the first caused data loss.

Fix this by recording a single pipeline barrier to cover every slice of
interest.

Acked-by: Rohan Garg <rohan.garg@intel.com>