func.mesh: Don't read from Mesh output variable in test
In VK_EXT_mesh_shader, mesh outputs can't be read, so don't use that on
test.

These tests came from the earlier VK_NV_mesh_shader tests, in which the
outputs could be read -- effectively acting as a shared memory for Mesh.
The original test takes advantage of that behavior by making some invocations
write to the outputs and all invocations read from it.

Reviewed-by: Ivan Briano <ivan.briano@intel.com>