Test bad optimization ordering that can rematerialize fsign
These next few lines look a bit like and open-coded sign(). It is
possible that optimization passes will convert things that look a bit
like an open-coded sign() to sign(). The problem occurs in the Intel
compiler when that optimization occurs late in compilation when sign()
instructions are not expected to exist. This can occur, per
mesa#12888, when 64-bit integer address arithmetic is lowered.

The sign() instruction is unknown to the backend. In the current code,
that case falls through to the case for trunc(). Input X coordinate is
chosen so that trunc(a_position.x) would not produce the same value as
sign(a_position.x).

The goofiness in the shader with the atomicAdd is necessary to end up
with this int64 address arithmetic on some Intel platforms.

v2: Fix mistakes adding the reference image. Noticed by Lionel. Rewrite
as a compute shader test using run_simple_compute_pipeline. Suggested by
Lionel.

Acked-by: Lionel Landwerlin <lionel.g.landwerlin@intel.com>
Ref: mesa#12888